Update zadatak2_vjezbe1.c
Ceil fix

+++ zadatak2_vjezbe1.c
@@ -1,13 +1,14 @@
 #include <stdio.h>
 // Zadatak 2, Vjezbe 1
+//TODO: Input with commas?
 
 int main(void) {
   float a, b;
   printf("Unesite duzine stranica pravougaonika (u formatu a, b): ");
   scanf("%f %f", &a, &b);
-  float povrsina = a * b;
-  float obim = (2 * a) + (2 * b);
-  printf("Obim pravougaonika iznosi %f, a povrsina %f", obim, povrsina);
+  int povrsina = a * b;
+  int obim = (2 * a) + (2 * b);
+  printf("Obim pravougaonika iznosi %d, a povrsina %d", obim, povrsina);
 
   return 0;
 }
